+++ pyledger/tests/__init__.py
@@ -5,7 +5,7 @@
 
 from .base_test_tax_codes import BaseTestTaxCodes
 from .base_test_accounts import BaseTestAccounts
-from .base_test_ledger import BaseTestJournal
+from .base_test_journal import BaseTestJournal
 from .base_test_assets import BaseTestAssets
 from .base_test_dump_restore_clear import BaseTestDumpRestoreClear
 from .base_test import BaseTest
